Move CachedData and HTTPSource classes into separate files
