+++ lib/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+"""Just init.
+
+I love how black and reorder-python-imports play nicely together and no
+workarounds are needed what so ever.
+"""
+from .cached_data import CachedData  # noqa: F401
+from .http_source import HTTPSource  # noqa: F401
