+++ lib/config_options.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+"""Common configuration options.
+
+I love how black and reorder-python-imports play nicely together and no
+workarounds are needed what so ever.
+"""
+CACHE_EXPIRATION = 86400  # seconds
+DATA_SOURCE_EXPIRATION = 30 * 86400  # seconds
+HTTP_TIMEOUT = 30  # seconds
